# Improve equals check