--- a/libbeat/conditions/equals.go
+++ b/libbeat/conditions/equals.go
@@ -80,8 +80,10 @@ func (c Equals) Check(event ValuesMap) bool {
 					return false
 				}
 
-				continue
+			} else {
+				logp.L().Named(logName).Warnf("expected int but got type %T in equals condition.", value)
 			}
+			continue
 
 		case string:
 			sValue, err := ExtractString(value)
@@ -89,9 +91,10 @@ func (c Equals) Check(event ValuesMap) bool {
 				if sValue != equalValue.Str {
 					return false
 				}
-
-				continue
+			} else {
+				logp.L().Named(logName).Warnf("expected string but got type %T in equals condition.", value)
 			}
+			continue
 
 		case bool:
 			bValue, err := ExtractBool(value)
@@ -99,9 +102,10 @@ func (c Equals) Check(event ValuesMap) bool {
 				if bValue != equalValue.Bool {
 					return false
 				}
-
-				continue
+			} else {
+				logp.L().Named(logName).Warnf("expected bool but got type %T in equals condition.", value)
 			}
+			continue
 		}
 
 		logp.L().Named(logName).Warnf("unexpected type %T in equals condition as it accepts only integers, strings, or booleans.", value)
@@ -113,4 +117,177 @@ func (c Equals) Check(event ValuesMap) bool {
 
 func (c Equals) String() string {
 	return fmt.Sprintf("equals: %v", map[string]equalsValue(c))
+}
+
+// Equals is a Condition for testing string equality.
+type Equals2 map[string]equalsValueType
+
+// equalValue checks its defined value equals the given value
+type equalsValueType interface {
+	Check(interface{}) bool
+}
+
+type equalsIntValue uint64
+
+func (e equalsIntValue) Check(value interface{}) bool {
+	if intValue, err := ExtractInt(value); err == nil {
+		return intValue == uint64(e)
+	}
+	logp.L().Named(logName).Warnf("expected int but got type %T in equals condition.", value)
+	return false
+}
+
+type equalsStringValue string
+
+func (e equalsStringValue) Check(value interface{}) bool {
+	if sValue, err := ExtractString(value); err == nil {
+		return sValue == string(e)
+	}
+	logp.L().Named(logName).Warnf("expected string but got type %T in equals condition.", value)
+	return false
+}
+
+type equalsBoolValue bool
+
+func (e equalsBoolValue) Check(value interface{}) bool {
+	if bValue, err := ExtractBool(value); err != nil {
+		return bValue == bool(e)
+	}
+	logp.L().Named(logName).Warnf("expected bool but got type %T in equals condition.", value)
+	return false
+}
+
+// NewEqualsCondition2 builds a new Equals using the given configuration of string equality checks.
+func NewEqualsCondition2(fields map[string]interface{}) (c Equals2, err error) {
+	c = Equals2{}
+
+	for field, value := range fields {
+		uintValue, err := ExtractInt(value)
+		if err == nil {
+			c[field] = equalsIntValue(uintValue)
+			continue
+		}
+
+		sValue, err := ExtractString(value)
+		if err == nil {
+			c[field] = equalsStringValue(sValue)
+			continue
+		}
+
+		bValue, err := ExtractBool(value)
+		if err == nil {
+			c[field] = equalsBoolValue(bValue)
+			continue
+		}
+
+		return nil, fmt.Errorf("condition attempted to set '%v' -> '%v' and encountered unexpected type '%T', only strings, ints, and booleans are allowed", field, value, value)
+	}
+
+	return c, nil
+}
+
+// Check determines whether the given event matches this condition.
+func (c Equals2) Check(event ValuesMap) bool {
+	for field, equalValue := range c {
+
+		value, err := event.GetValue(field)
+		if err != nil {
+			return false
+		}
+
+		if !equalValue.Check(value) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func (c Equals2) String() string {
+	return fmt.Sprintf("equals: %v", map[string]equalsValueType(c))
+}
+
+// Equals2 is a Condition for testing string equality.
+type Equals3 map[string]equalsValueFunc
+
+type equalsValueFunc func(interface{}) bool
+
+func equalsIntValue3(i uint64) equalsValueFunc {
+	return func(value interface{}) bool {
+		if sValue, err := ExtractInt(value); err == nil {
+			return sValue == i
+		}
+		logp.L().Named(logName).Warnf("expected int but got type %T in equals condition.", value)
+		return false
+	}
+}
+
+func equalsStringValue3(s string) equalsValueFunc {
+	return func(value interface{}) bool {
+		if sValue, err := ExtractString(value); err == nil {
+			return sValue == s
+		}
+		logp.L().Named(logName).Warnf("expected string but got type %T in equals condition.", value)
+		return false
+	}
+}
+
+func equalsBoolValue3(b bool) equalsValueFunc {
+	return func(value interface{}) bool {
+		if sValue, err := ExtractBool(value); err == nil {
+			return sValue == b
+		}
+		logp.L().Named(logName).Warnf("expected bool but got type %T in equals condition.", value)
+		return false
+	}
+}
+
+// NewEqualsCondition3 builds a new Equals using the given configuration of string equality checks.
+func NewEqualsCondition3(fields map[string]interface{}) (c Equals3, err error) {
+	c = Equals3{}
+
+	for field, value := range fields {
+		uintValue, err := ExtractInt(value)
+		if err == nil {
+			c[field] = equalsIntValue3(uintValue)
+			continue
+		}
+
+		sValue, err := ExtractString(value)
+		if err == nil {
+			c[field] = equalsStringValue3(sValue)
+			continue
+		}
+
+		bValue, err := ExtractBool(value)
+		if err == nil {
+			c[field] = equalsBoolValue3(bValue)
+			continue
+		}
+
+		return nil, fmt.Errorf("condition attempted to set '%v' -> '%v' and encountered unexpected type '%T', only strings, ints, and booleans are allowed", field, value, value)
+	}
+
+	return c, nil
+}
+
+// Check determines whether the given event matches this condition.
+func (c Equals3) Check(event ValuesMap) bool {
+	for field, equalValue := range c {
+
+		value, err := event.GetValue(field)
+		if err != nil {
+			return false
+		}
+
+		if !equalValue(value) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func (c Equals3) String() string {
+	return fmt.Sprintf("equals: %v", map[string]equalsValueFunc(c))
 }