## What does this PR do?

Improve equals check performance

## Why is it important?

* Improve equals check performance
* improve log output
* fix potentially wrong matches

## Checklist

- [x] My code follows the style guidelines of this project
- [x] ~~I have commented my code, particularly in hard-to-understand areas~~
- [x] ~~I have made corresponding changes to the documentation~~
- [x] ~~I have made corresponding change to the default configuration files~~
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.

## Author's Checklist

- [ ]

## How to test this PR locally

Benchmark test

## Logs

Benchmark added test case, before this change
```
pkg: github.com/elastic/beats/v7/libbeat/conditions
BenchmarkEquals_Check
BenchmarkEquals_Check-8   	 1529768	       829 ns/op
PASS
```

After
```
pkg: github.com/elastic/beats/v7/libbeat/conditions
BenchmarkEquals_Check
BenchmarkEquals_Check-8   	 6298933	       188 ns/op
PASS
```