@urso 2020-12-02
Nice!!!

Have you considered to turn `equalsValue` into an interface and do something like this?

```
type equalsValue interface {
  Check(ValuesMap) bool
}

type equalsIntValue int
type equalsStringValue string
...

func (v equalsIntValue) Check(event ValuesMap) bool {
   ...
}
```

alternatively one can try functions as well like:

```
type equalsValue func(ValuesMap) bool

func equalsIntValue(i int) equalsValue {
  return func(event ValuesMap) bool {
    ...
  }
}
```

equalsValue is internal, so we can play with it. Just wondering which implementation would be the fastest here.

All in all I would prefer to move the loop body into a method on `equalsValue`.