@newly12 2020-12-02
Thanks for the review. Both look to be more elegant ways to me.

I was trying to implement this and notice that it seems that in both implementation `equalsValue` has to take an `interface{}` as its parameter instead of `ValuesMap`, because the name of the field that is used to extract the value is still saved in `Equals` type. Please correct me if I am wrong, and let me know if I shall move on with the interface parameter.

```
type Equals map[string]equalsValue
```

Another notice is that warning logs will not be logged anymore after this change, since for a field `Equals.check` does not try to extract all supported types from the value but only its wanted type, FYI in case this concerns you.