@urso 2020-12-02
> Another notice is that warning logs will not be logged anymore after this change, since for a field Equals.check does not try to extract all supported types from the value but only its wanted type, FYI in case this concerns you.

Yes, this is a concern. If checks do not work as expected for all events due to type mismatches, then the logs will point this out.

> I was trying to implement this and notice that it seems that in both implementation equalsValue has to take an interface{} as its parameter instead of ValuesMap, because the name of the field that is used to extract the value is still saved in Equals type.

Ah, you are right. Didn't notice that value is an interface. In that case we would have `type equalsValue interface { Check(value interface{}) (bool, error) }`, right? Then we can log the error if the extraction failed.