@newly12 2020-12-03
> Ah, you are right. Didn't notice that value is an interface. In that case we would have type equalsValue interface { Check(value interface{}) (bool, error) }, right? Then we can log the error if the extraction failed.

yeah, I think the difference would be 1. previously warning logs is about all type are mismatched 2. now it warns if the value is not its wanted type. For example, if I have a config `"strField": "foo"`, I got a event with `"strField": 1`, previously there is no warning log, because the value will be extracted as int, though it doesn't equal "foo", now there will be a warning, we expect string but we have a int extracted. 

The result is still the same though.