@urso 2020-12-03
> now it warns if the value is not its wanted type. For example, if I have a config "strField": "foo", I got a event with "strField": 1, previously there is no warning log, because the value will be extracted as int, though it doesn't equal "foo", now there will be a warning, we expect string but we have a int extracted.

I would say this is an improvement :)

Plus the old code did not take the type in the config in mind when comparing the results. E.g. if `ExtractInt` on the event did succeed, it would have compared the result with the integer value `0`, if the actual setting in the config was a string. This could result in wrong true/false results.

The change will: improve performance, improve log output, fix potentially wrong matches.