@newly12 2020-12-04
Hi,

I've added both implementation in the code for now for your reference, the benchmark result is attached below, overall it seems the implementation with new type wrapping int/str/bool has better performance, but I am a bit confused why it shows a lower than others when there are less checks, which I couldn't figure out at this moment..

```
BenchmarkEqualsWith1Conditions
BenchmarkEqualsWith1Conditions-8    	14561558	        77.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkEquals2With1Conditions
BenchmarkEquals2With1Conditions-8   	13009749	        85.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkEquals3With1Conditions
BenchmarkEquals3With1Conditions-8   	14077786	        79.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkEqualsWith3Conditions
BenchmarkEqualsWith3Conditions-8    	 5666799	       205 ns/op	       0 B/op	       0 allocs/op
BenchmarkEquals2With3Conditions
BenchmarkEquals2With3Conditions-8   	 3171140	       384 ns/op	     216 B/op	       4 allocs/op
BenchmarkEquals3With3Conditions
BenchmarkEquals3With3Conditions-8   	 6365893	       195 ns/op	       0 B/op	       0 allocs/op
BenchmarkEqualsWith5Conditions
BenchmarkEqualsWith5Conditions-8    	 2231367	       543 ns/op	     245 B/op	       5 allocs/op
BenchmarkEquals2With5Conditions
BenchmarkEquals2With5Conditions-8   	 3334934	       337 ns/op	     197 B/op	       3 allocs/op
BenchmarkEquals3With5Conditions
BenchmarkEquals3With5Conditions-8   	 3765338	       470 ns/op	     245 B/op	       5 allocs/op
BenchmarkEqualsWith7Conditions
BenchmarkEqualsWith7Conditions-8    	 1584600	       674 ns/op	     315 B/op	       6 allocs/op
BenchmarkEquals2With7Conditions
BenchmarkEquals2With7Conditions-8   	 3286808	       404 ns/op	     221 B/op	       4 allocs/op
BenchmarkEquals3With7Conditions
BenchmarkEquals3With7Conditions-8   	 2253856	       486 ns/op	     245 B/op	       5 allocs/op
```