@newly12 2020-12-08
Thanks @urso . I had compared twice since the result from some other benchmark seems confusing or misleading, the result for equals check is consistent.  

```
$ benchcmp old.txt new.txt   
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                          old ns/op     new ns/op     delta
BenchmarkSimpleCondition-8         47.6          24.7          -48.11%
BenchmarkCombinedCondition-8       134           221           +64.93%
BenchmarkEquals/1_condition-8      336           79.3          -76.40%
BenchmarkEquals/3_conditions-8     824           203           -75.36%
BenchmarkEquals/5_conditions-8     1285          315           -75.49%
BenchmarkEquals/7_conditions-8     917           354           -61.40%
BenchmarkNetworkCondition-8        173           171           -1.16%

$ benchcmp old2.txt new2.txt 
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                          old ns/op     new ns/op     delta
BenchmarkSimpleCondition-8         22.5          35.6          +58.22%
BenchmarkCombinedCondition-8       136           141           +3.68%
BenchmarkEquals/1_condition-8      432           73.6          -82.96%
BenchmarkEquals/3_conditions-8     786           194           -75.32%
BenchmarkEquals/5_conditions-8     1224          315           -74.26%
BenchmarkEquals/7_conditions-8     842           278           -66.98%
BenchmarkNetworkCondition-8        165           180           +9.09%
```